pyproject.toml
[project]
name = "lab-github-cicd-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "flask>=3.0.0",
    "gunicorn>=22.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.2",
    "sentry-sdk[flask]>=2.40.0",
]

[dependency-groups]
dev = [
    "pylint>=3.2.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
    "tests",
]
